+++ src/App.jsx
@@ -1,5 +1,5 @@
 import ProductsDisplay from "./components/ProductsDisplay";
-import Navbar from "./components/Navbar";
+import Navbar from "./components/navbar/Navbar";
 import { useState } from "react";
 
 function App() {
